PyQgis: Move NULL and edit to qgis.core
